用户登录日志添加

+++ src/main/java/org/ewlive/entity/system/SysLogOperate.java
@@ -44,7 +44,7 @@ public class SysLogOperate extends Base {
      * 操作耗时(ms)
      */
     @TableField("task_time_span")
-    private Integer taskTimeSpan;
+    private Long taskTimeSpan;
 
 
     /**
